+++ src/main/java/keqing/pollution/common/metatileentity/multiblock/MetaTileEntityTankTest.java
@@ -1,17 +1,18 @@
 package keqing.pollution.common.metatileentity.multiblock;
 
 import gregicality.multiblocks.api.render.GCYMTextures;
-import gregicality.multiblocks.common.block.GCYMMetaBlocks;
-import gregicality.multiblocks.common.block.blocks.BlockLargeMultiblockCasing;
-import gregtech.api.capability.GregtechDataCodes;
+import gregtech.api.capability.IMultipleTankHandler;
+import gregtech.api.capability.impl.FluidTankList;
 import gregtech.api.metatileentity.IFastRenderMetaTileEntity;
 import gregtech.api.metatileentity.MetaTileEntity;
 import gregtech.api.metatileentity.interfaces.IGregTechTileEntity;
 import gregtech.api.metatileentity.multiblock.IMultiblockPart;
+import gregtech.api.metatileentity.multiblock.MultiblockAbility;
+import gregtech.api.metatileentity.multiblock.MultiblockWithDisplayBase;
 import gregtech.api.pattern.BlockPattern;
 import gregtech.api.pattern.FactoryBlockPattern;
-import gregtech.api.recipes.RecipeMap;
-import gregtech.api.recipes.RecipeMaps;
+import gregtech.api.pattern.PatternMatchContext;
+import gregtech.api.util.GTTransferUtils;
 import gregtech.api.util.RelativeDirection;
 import gregtech.api.util.interpolate.Eases;
 import gregtech.client.renderer.ICubeRenderer;
@@ -25,35 +26,183 @@ import gregtech.common.blocks.BlockGlassCasing;
 import gregtech.common.blocks.BlockTurbineCasing;
 import gregtech.common.blocks.MetaBlocks;
 import keqing.pollution.api.metatileentity.POMultiblockAbility;
-import keqing.pollution.api.metatileentity.PORecipeMapMultiblockController;
-import keqing.pollution.api.metatileentity.POTankMultiblockController;
+import keqing.pollution.api.unification.PollutionMaterials;
+import keqing.pollution.client.textures.POTextures;
+import keqing.pollution.common.block.PollutionMetaBlock.POMagicBlock;
+import keqing.pollution.common.block.PollutionMetaBlocks;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.client.renderer.BufferBuilder;
 import net.minecraft.client.renderer.GlStateManager;
 import net.minecraft.client.renderer.OpenGlHelper;
 import net.minecraft.client.renderer.Tessellator;
 import net.minecraft.client.renderer.vertex.DefaultVertexFormats;
+import net.minecraft.nbt.NBTTagCompound;
+import net.minecraft.network.PacketBuffer;
 import net.minecraft.util.EnumFacing;
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.math.AxisAlignedBB;
+import net.minecraft.util.text.ITextComponent;
+import net.minecraft.util.text.TextComponentTranslation;
+import net.minecraftforge.fluids.FluidStack;
+import net.minecraftforge.fluids.IFluidTank;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
 import org.lwjgl.opengl.GL11;
+import thaumcraft.api.aspects.Aspect;
 import thaumcraft.common.tiles.essentia.TileJarFillable;
 
-public class MetaTileEntityTankTest extends POTankMultiblockController implements IBloomEffect , IFastRenderMetaTileEntity {
+import java.util.Collections;
+import java.util.List;
+import java.util.Objects;
 
+public class MetaTileEntityTankTest extends MultiblockWithDisplayBase implements IBloomEffect , IFastRenderMetaTileEntity {
+    int aX = 0;
+    int aY = 0;
+    int aZ = 0;
+    public Aspect al;
+    public String name = null;
+    public int storage = 0;
+    public int number = 0;
+    FluidStack AIR_STACK;
+    FluidStack FIRE_STACK;
+    FluidStack WATER_STACK;
+    FluidStack ERATH_STACK;
+    FluidStack ORDER_STACK;
+    FluidStack ENTROPY_STACK;
     int RadomTime=0;
+    protected IMultipleTankHandler outputFluidInventory;
     public MetaTileEntityTankTest(ResourceLocation metaTileEntityId) {
-        super(metaTileEntityId, new RecipeMap[] { RecipeMaps.EXTRACTOR_RECIPES, RecipeMaps.COMPRESSOR_RECIPES,
-                RecipeMaps.FORMING_PRESS_RECIPES, RecipeMaps.FORGE_HAMMER_RECIPES });
+        super(metaTileEntityId);
+        this.AIR_STACK = PollutionMaterials.infused_air.getFluid(1);
+        this.FIRE_STACK = PollutionMaterials.infused_fire.getFluid(1);
+        this.WATER_STACK = PollutionMaterials.infused_water.getFluid(1);
+        this.ERATH_STACK = PollutionMaterials.infused_earth.getFluid(1);
+        this.ORDER_STACK = PollutionMaterials.infused_order.getFluid(1);
+        this.ENTROPY_STACK = PollutionMaterials.infused_entropy.getFluid(1);
+    }
+    protected void initializeAbilities() {
+        this.outputFluidInventory = new FluidTankList(this.allowSameFluidFillForOutputs(), this.getAbilities(MultiblockAbility.EXPORT_FLUIDS));
+    }
+    protected boolean allowSameFluidFillForOutputs() {
+        return true;
+    }
+    private void resetTileAbilities() {
+        this.outputFluidInventory = new FluidTankList(true, new IFluidTank[0]);
+    }
+
+    public void invalidateStructure() {
+        super.invalidateStructure();
+        this.resetTileAbilities();
     }
 
     @Override
-    public MetaTileEntity createMetaTileEntity(IGregTechTileEntity metaTileEntityHolder) {
+    public MetaTileEntity createMetaTileEntity(IGregTechTileEntity iGregTechTileEntity) {
         return new MetaTileEntityTankTest(this.metaTileEntityId);
     }
+    public boolean fillTanks(FluidStack stack, boolean simulate) {
+        return GTTransferUtils.addFluidsToFluidHandler(this.getOutputFluidInventory(), simulate, Collections.singletonList(stack));
+    }
+    public IMultipleTankHandler getOutputFluidInventory() {
+        return this.outputFluidInventory;
+    }
+    public void updateFormedValid() {
+        super.update();
+        if (this.getWorld().getTileEntity(this.getPos().add(0, 2, 0)) instanceof TileJarFillable) {
+            TileJarFillable s = (TileJarFillable)this.getWorld().getTileEntity(this.getPos().add(0, 2, 0));
+            this.al = s.getEssentiaType(s.getFacing());
+            this.storage = s.getEssentiaAmount(s.getFacing());
+            if (this.al != null) {
+                this.name = this.al.getName();
+            }
+        }
 
+        if (this.number == 0) {
+            this.number = this.storage;
+            this.clearInfused();
+            this.name = null;
+        }
+
+        this.doFillTank();
+
+        if(RadomTime<10)RadomTime++;
+        else RadomTime=0;
+        setFusionRingColor(0xFF000000+RadomTime*100);
+    }
+
+    private void doFillTank() {
+        if (Objects.equals(this.name, "Aer") && this.number > 0) {
+            this.fillTanks(this.AIR_STACK, false);
+            --this.number;
+        }
+
+        if (Objects.equals(this.name, "Terra") && this.number > 0) {
+            this.fillTanks(this.ERATH_STACK, false);
+            --this.number;
+        }
+
+        if (Objects.equals(this.name, "Aqua") && this.number > 0) {
+            this.fillTanks(this.WATER_STACK, false);
+            --this.number;
+        }
+
+        if (Objects.equals(this.name, "Ignis") && this.number > 0) {
+            this.fillTanks(this.FIRE_STACK, false);
+            --this.number;
+        }
+
+        if (Objects.equals(this.name, "Ordo") && this.number > 0) {
+            this.fillTanks(this.ORDER_STACK, false);
+            --this.number;
+        }
+
+        if (Objects.equals(this.name, "Perdition") && this.number > 0) {
+            this.fillTanks(this.ENTROPY_STACK, false);
+            --this.number;
+        }
+
+    }
+
+    public void clearInfused() {
+        if (this.getWorld().getTileEntity(this.getPos().add(0, 2, 0)) instanceof TileJarFillable) {
+            TileJarFillable s = (TileJarFillable)this.getWorld().getTileEntity(this.getPos().add(0, 2, 0));
+            s.amount = 0;
+        }
+
+    }
+
+    protected void formStructure(PatternMatchContext context) {
+        super.formStructure(context);
+        this.initializeAbilities();
+    }
+
+    protected void addDisplayText(List<ITextComponent> textList) {
+        super.addDisplayText(textList);
+        if (this.isStructureFormed()) {
+            textList.add(new TextComponentTranslation("Infused: %s  Amount: %s", new Object[]{this.name, this.storage}));
+        }
+
+        textList.add(new TextComponentTranslation("Infused: %s  Amount: %s", new Object[]{this.name, this.number}));
+    }
+
+    public NBTTagCompound writeToNBT(NBTTagCompound data) {
+        data.setInteger("number", this.number);
+        return super.writeToNBT(data);
+    }
+
+    public void readFromNBT(NBTTagCompound data) {
+        super.readFromNBT(data);
+        this.number = data.getInteger("number");
+    }
+
+    public void writeInitialSyncData(PacketBuffer buf) {
+        super.writeInitialSyncData(buf);
+        buf.writeInt(this.number);
+    }
+
+    public void receiveInitialSyncData(PacketBuffer buf) {
+        super.receiveInitialSyncData(buf);
+        this.number = buf.readInt();
+    }
     @Override
     protected  BlockPattern createStructurePattern() {
         return FactoryBlockPattern.start()
@@ -70,8 +219,8 @@ public class MetaTileEntityTankTest extends POTankMultiblockController implement
     }
 
     private static IBlockState getCasingState() {
-        return GCYMMetaBlocks.LARGE_MULTIBLOCK_CASING
-                .getState(BlockLargeMultiblockCasing.CasingType.STRESS_PROOF_CASING);
+        return PollutionMetaBlocks.MAGIC_BLOCK
+                .getState(POMagicBlock.MagicBlockType.VOID_PRISM);
     }
 
     private static IBlockState getCasingState2() {
@@ -84,32 +233,13 @@ public class MetaTileEntityTankTest extends POTankMultiblockController implement
 
     @Override
     public ICubeRenderer getBaseTexture(IMultiblockPart iMultiblockPart) {
-        return GCYMTextures.STRESS_PROOF_CASING;
+        return POTextures.VOID_PRISM;
     }
 
     @Override
     protected  OrientedOverlayRenderer getFrontOverlay() {
         return GCYMTextures.LARGE_BENDER_OVERLAY;
     }
-
-    @Override
-    public boolean canBeDistinct() {
-        return true;
-    }
-    @Override
-    public void update() {
-        super.update();
-        if(RadomTime<10)RadomTime++;
-        else RadomTime=0;
-        setFusionRingColor(0xFF000000+RadomTime*100);
-
-    }
-
-
-
-
-
-
 
     protected static final int NO_COLOR = 0;
     private int fusionRingColor = NO_COLOR;
@@ -130,7 +260,6 @@ public class MetaTileEntityTankTest extends POTankMultiblockController implement
             //writeCustomData(GregtechDataCodes.UPDATE_COLOR, buf -> buf.writeVarInt(fusionRingColor));
         }
     }
-
 
     @Override
     @SideOnly(Side.CLIENT)
