+++ src/main/resources/assets/gregtech/models/item/metaitems/coking_catalyst_core.json
@@ -1,6 +1,6 @@
 {
   "parent": "item/generated",
   "textures": {
-    "layer0": "pollution:items/metaitems/test"
+    "layer0": "pollution:items/metaitems/coking_catalyst_core"
   }
 }
