+++ src/main/resources/assets/gregtech/models/item/metaitems/glue_slime.json
@@ -1,6 +1,6 @@
 {
   "parent": "item/generated",
   "textures": {
-    "layer0": "pollution:items/metaitems/test"
+    "layer0": "pollution:items/metaitems/glue_slime"
   }
 }
