+++ src/main/resources/assets/gregtech/models/item/metaitems/sugar_slime.json
@@ -1,0 +1,6 @@
+{
+  "parent": "item/generated",
+  "textures": {
+    "layer0": "pollution:items/metaitems/sugar_slime"
+  }
+}
