+++ src/main/resources/assets/gregtech/models/item/metaitems/tar_slime.json
@@ -1,6 +1,6 @@
 {
   "parent": "item/generated",
   "textures": {
-    "layer0": "pollution:items/metaitems/test"
+    "layer0": "pollution:items/metaitems/tar_slime"
   }
 }
