+++ src/main/resources/assets/pollution/textures/items/metaitems/coking_catalyst_core.png.mcmeta
@@ -1,0 +1,1 @@
+{"animation": {"frametime": 1}}
